collect some journals
we get a list of all units by calling systemd's ListUnits dbus api, and
iterate through them to find out which units were loaded from
/usr/lib/systemd/system, and then ask journalctl for their logs.

we collect just these journals so that hopefully we will not collect
logs from services created by users which may put private information
into the journal.

this method is not very good, but should suffice for now. in the future,
a whitelist of services should probably be used.